requirements.txt
google-genai
clip-interrogator
openai-clip
pillow
requests
tqdm
aiohttp
aiofiles
faiss-cpu
python-dotenv
torch
numpy
scikit-learn
gradio
ultralytics